python/requirements.txt
face_recognition>=1.3.0
Pillow>=9.0
tqdm>=4.60
numpy>=1.20
